# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Pulse Healthcare API",
        "description": "A hospital booking and management platform connecting patients with healthcare facilities in Ghana. Manages hospital registration, doctor scheduling, appointment booking, and real-time queue tracking.",
        "contact": {
            "name": "Marvinphil Annorbah",
            "url": "https://github.com/psam-717/pulse",
            "email": "mphilannorbah@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "v2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearer-jwt": []
        }
    ],
    "paths": {
        "/api/hospitals/{hospitalId}/working-hours": {
            "get": {
                "tags": [
                    "hospital-controller"
                ],
                "operationId": "getWorkingHours",
                "parameters": [
                    {
                        "name": "hospitalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WorkingHours"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": []
            },
            "put": {
                "tags": [
                    "hospital-controller"
                ],
                "operationId": "configureWorkingHours",
                "parameters": [
                    {
                        "name": "hospitalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkingHoursRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WorkingHours"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/hospitals/{hospitalId}/license": {
            "get": {
                "tags": [
                    "hospital-controller"
                ],
                "operationId": "getLicenseInfo",
                "parameters": [
                    {
                        "name": "hospitalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/HospitalResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "hospital-controller"
                ],
                "operationId": "uploadLicense",
                "parameters": [
                    {
                        "name": "hospitalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/HospitalResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/hospitals/{hospitalId}/departments": {
            "get": {
                "tags": [
                    "department-controller"
                ],
                "operationId": "listDepartments",
                "parameters": [
                    {
                        "name": "hospitalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Department"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": []
            },
            "post": {
                "tags": [
                    "hospital-controller"
                ],
                "operationId": "createDepartment",
                "parameters": [
                    {
                        "name": "hospitalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepartmentRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Department"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/hospitals/register": {
            "post": {
                "tags": [
                    "hospital-controller"
                ],
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HospitalRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegistrationResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/hospitals/login": {
            "post": {
                "tags": [
                    "hospital-controller"
                ],
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HospitalLoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
        "/api/hospitals/{hospitalId}": {
            "get": {
                "tags": [
                    "hospital-controller"
                ],
                "operationId": "getHospital",
                "parameters": [
                    {
                        "name": "hospitalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/HospitalResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
        "/api/doctors/{doctorId}/slots": {
            "get": {
                "tags": [
                    "time-slot-controller"
                ],
                "operationId": "listAvailableSlots",
                "parameters": [
                    {
                        "name": "doctorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TimeSlot"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
        "/api/hospitals/{hospitalId}/departments/{departmentId}": {
            "delete": {
                "tags": [
                    "hospital-controller"
                ],
                "operationId": "deleteDepartment",
                "parameters": [
                    {
                        "name": "hospitalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "departmentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "WorkingHoursEntry": {
                "type": "object",
                "properties": {
                    "dayOfWeek": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "openTime": {
                        "type": "string"
                    },
                    "closeTime": {
                        "type": "string"
                    },
                    "isClosed": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "closeTime",
                    "dayOfWeek",
                    "openTime"
                ]
            },
            "WorkingHoursRequest": {
                "type": "object",
                "properties": {
                    "entries": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/WorkingHoursEntry"
                        }
                    }
                }
            },
            "Hospital": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "licenseNumber": {
                        "type": "string"
                    },
                    "licenseDocumentUrl": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "specialties": {
                        "type": "string"
                    },
                    "capacity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "verificationStatus": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "APPROVED",
                            "REJECTED"
                        ]
                    },
                    "rejectionReason": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "WorkingHours": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "hospital": {
                        "$ref": "#/components/schemas/Hospital"
                    },
                    "dayOfWeek": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "openTime": {
                        "type": "string"
                    },
                    "closeTime": {
                        "type": "string"
                    },
                    "closed": {
                        "type": "boolean"
                    }
                }
            },
...
            "HospitalResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "licenseNumber": {
                        "type": "string"
                    },
                    "licenseDocumentUrl": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "specialties": {
                        "type": "string"
                    },
                    "capacity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "verificationStatus": {
                        "type": "string"
                    },
                    "rejectionReason": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    }
                }
            },
            "DepartmentRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1
                    },
                    "abbreviation": {
                        "type": "string",
                        "minLength": 1
                    },
                    "description": {
                        "type": "string"
                    },
                    "consultationFee": {
                        "type": "number"
                    },
                    "parentDepartmentId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "abbreviation",
                    "consultationFee",
                    "name"
                ]
            },
            "Department": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "abbreviation": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "consultationFee": {
                        "type": "number"
                    },
                    "parentDepartment": {
                        "$ref": "#/components/schemas/Department"
                    },
                    "hospital": {
                        "$ref": "#/components/schemas/Hospital"
                    }
                }
            },
            "HospitalRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1
                    },
                    "licenseNumber": {
                        "type": "string",
                        "minLength": 1
                    },
                    "licenseDocumentUrl": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string",
                        "minLength": 1
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "specialties": {
                        "type": "string"
                    },
                    "capacity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "phone": {
                        "type": "string",
                        "minLength": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "adminFullName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "adminEmail": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "adminPassword": {
                        "type": "string",
                        "maxLength": 2147483647,
                        "minLength": 6
                    },
                    "adminPhone": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "address",
                    "adminEmail",
                    "adminFullName",
                    "adminPassword",
                    "adminPhone",
                    "email",
                    "licenseNumber",
                    "name",
                    "phone"
                ]
            },
            "RegistrationResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "hospitalId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "adminId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "token": {
                        "type": "string"
                    }
                }
            },
            "HospitalLoginRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "AuthResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "ApiResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "Doctor": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "specialization": {
                        "type": "string"
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "licenseNumber": {
                        "type": "string"
                    },
                    "workspaceId": {
                        "type": "string"
                    },
                    "consultationDuration": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "hospital": {
                        "$ref": "#/components/schemas/Hospital"
                    }
                }
            },
            "TimeSlot": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "doctor": {
                        "$ref": "#/components/schemas/Doctor"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "startTime": {
                        "type": "string"
                    },
                    "endTime": {
                        "type": "string"
                    },
                    "booked": {
                        "type": "boolean"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer-jwt": {
                "type": "http",
                "description": "Enter your JWT token. Get one from POST /api/auth/login or /api/auth/register\n\nExample: admin token \u2192 `Bearer eyJhbGciOiJIUzI1NiJ...`",
                "name": "bearer-jwt",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}